DigitalShadows(PAPP-11277): Added the python 2to3 conversion

+++ Apps/phdigitalshadows/dsapi/model/data_breach.py
@@ -99,18 +99,18 @@ class DataBreach(DSModel):
     @classmethod
     def from_json(cls, json):
         cast = DSModel.cast
-        return cls(cast(json.get('id'), long),
+        return cls(cast(json.get('id'), int),
                    json.get('title'),
                    json.get('domainName'),
                    json.get('occurred'),
                    json.get('modified'),
                    json.get('published'),
-                   cast(json.get('incident').get('id'), long),
+                   cast(json.get('incident').get('id'), int),
                    json.get('incident').get('scope'),
                    json.get('incident').get('type'),
                    json.get('incident').get('severity'),
                    json.get('incident').get('title'),
-                   cast(json.get('domainCount'), long),
-                   cast(json.get('recordCount'), long),
+                   cast(json.get('domainCount'), int),
+                   cast(json.get('recordCount'), int),
                    json.get('sourceUrl'),
-                   cast(json.get('organisationUsernameCount'), long), json)
+                   cast(json.get('organisationUsernameCount'), int), json)
